# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/posts": {
      "get": {
        "description": "게시글 목록을 조회합니다. 검색, 필터링, 페이징 기능을 제공합니다.",
        "operationId": "PostsController_findAll",
        "parameters": [
          {
            "name": "searchTerm",
            "required": false,
            "in": "query",
            "description": "검색어 (제목, 내용)",
            "schema": {
              "example": "검색어",
              "type": "string"
            }
          },
          {
            "name": "tags",
            "required": false,
            "in": "query",
            "description": "태그 필터 (쉼표로 구분 가능)",
            "schema": {
              "example": [
                "태그1",
                "태그2"
              ],
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "pageIndex",
            "required": false,
            "in": "query",
            "description": "페이지 인덱스 (0부터 시작)",
            "schema": {
              "default": 0,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "페이지 크기",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "게시글 목록 및 전체 개수를 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 파라미터입니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "인증에 실패했습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "권한이 없습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "리소스를 찾을 수 없습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "게시글 목록 조회",
        "tags": [
          "posts"
        ]
      },
      "post": {
        "description": "새로운 게시글을 생성합니다. 썸네일 이미지 업로드 가능.",
        "operationId": "PostsController_createPost",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "게시글 제목",
                    "example": "안녕하세요, 첫 번째 게시글입니다"
                  },
                  "content": {
                    "type": "string",
                    "description": "게시글 내용 (마크다운 지원)",
                    "example": "# 제목\n\n본문 내용은 마크다운을 지원합니다.\n\n- 목록 항목 1\n- 목록 항목 2"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "게시글 태그 목록",
                    "example": [
                      "태그1",
                      "태그2",
                      "태그3"
                    ]
                  },
                  "thumbnail": {
                    "type": "string",
                    "format": "binary",
                    "description": "썸네일 이미지 파일 (jpg, jpeg, png, webp, gif)"
                  }
                },
                "required": [
                  "title",
                  "content"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "생성된 게시글 정보를 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "400": {
            "description": "게시글 생성에 실패했습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "인증에 실패했습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "권한이 없습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "리소스를 찾을 수 없습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "게시글 생성",
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}": {
      "get": {
        "description": "ID로 특정 게시글을 조회합니다.",
        "operationId": "PostsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "게시글 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "게시글 정보를 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청입니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "인증에 실패했습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "권한이 없습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "게시글을 찾을 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "게시글 조회",
        "tags": [
          "posts"
        ]
      },
      "put": {
        "description": "기존 게시글을 수정합니다. 썸네일 이미지도 변경 가능합니다.",
        "operationId": "PostsController_updatePost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "수정할 게시글 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정된 게시글 정보를 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "400": {
            "description": "게시글 수정에 실패했습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "인증에 실패했습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "권한이 없습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "게시글을 찾을 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "게시글 수정",
        "tags": [
          "posts"
        ]
      },
      "delete": {
        "description": "게시글과 연관된 이미지를 삭제합니다.",
        "operationId": "PostsController_deletePost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "게시글 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "게시글이 성공적으로 삭제되었습니다."
          },
          "400": {
            "description": "게시글 삭제에 실패했습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "인증에 실패했습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "권한이 없습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "게시글을 찾을 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "게시글 삭제",
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/upload-image": {
      "post": {
        "description": "게시글 에디터에서 사용할 이미지를 업로드합니다. 허용된 이미지 타입: jpg, png, webp, gif, jpeg",
        "operationId": "PostsController_uploadContentImage",
        "parameters": [],
        "responses": {
          "201": {
            "description": "이미지 업로드 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string",
                      "description": "업로드된 이미지 URL",
                      "example": "https://res.cloudinary.com/demo/image/upload/v1123456789/post_content_images/sample.jpg"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "이미지 업로드 실패 또는 유효하지 않은 이미지 형식",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "인증에 실패했습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "권한이 없습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "리소스를 찾을 수 없습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "에디터 이미지 업로드",
        "tags": [
          "posts"
        ]
      }
    },
...
    "/comments/post/{postId}": {
      "get": {
        "description": "특정 게시글에 달린 모든 댓글을 조회합니다.",
        "operationId": "CommentsController_findAllByPostId",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "게시글 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "댓글 목록 반환",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentResponseDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "게시글을 찾을 수 없습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "게시글의 모든 댓글 조회",
        "tags": [
          "comments"
        ]
      },
      "post": {
        "description": "게시글에 댓글을 작성합니다. 로그인이 필요합니다.",
        "operationId": "CommentsController_create",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "게시글 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "댓글 작성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "댓글을 생성할 수 없습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "인증에 실패했습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "권한이 없습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "게시글을 찾을 수 없습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "댓글 작성",
        "tags": [
          "comments"
        ]
      }
    },
...
  },
  "info": {
    "title": "COCOMOMO API",
    "description": "COCOMOMO 백엔드 API 문서",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "에러 코드",
            "example": 400
          },
          "message": {
            "type": "string",
            "description": "에러 메시지",
            "example": "잘못된 요청입니다"
          },
          "error": {
            "type": "string",
            "description": "에러 타입",
            "example": "BadRequestException"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
      "Post": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "게시글 ID (자동생성)",
            "example": 1
          },
          "title": {
            "type": "string",
            "description": "게시글 제목",
            "example": "안녕하세요 첫 게시글입니다"
          },
          "content": {
            "type": "string",
            "description": "게시글 본문 내용 (마크다운 지원)",
            "example": "이것은 게시글의 본문 내용입니다. **마크다운**을 지원하며 이미지를 포함할 수 있습니다."
          },
          "thumbnailUrl": {
            "type": "string",
            "description": "썸네일 이미지 URL",
            "example": "https://res.cloudinary.com/demo/image/upload/v1123456789/sample.jpg"
          },
          "tags": {
            "description": "게시글 태그 목록",
            "example": [
              "일상",
              "여행",
              "취미"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "게시글 생성 시간",
            "example": "2025-06-09T13:30:51+09:00"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "게시글 수정 시간",
            "example": "2025-06-09T13:30:51+09:00"
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "createdAt",
          "updatedAt"
        ]
      },
      "PostsResponseDto": {
        "type": "object",
        "properties": {
          "list": {
            "description": "게시글 목록",
            "items": {
              "$ref": "#/components/schemas/Post"
            },
            "type": "array"
          },
          "totalCount": {
            "type": "number",
            "description": "전체 게시글 수",
            "example": 100
          }
        },
        "required": [
          "list",
          "totalCount"
        ]
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "게시글 제목",
            "example": "수정된 게시글 제목입니다"
          },
          "content": {
            "type": "string",
            "description": "게시글 본문 내용",
            "example": "이것은 수정된 게시글의 본문 내용입니다."
          },
          "tags": {
            "description": "게시글 태그 목록",
            "example": [
              "수정됨",
              "업데이트"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "thumbnailUrl": {
            "type": "string",
            "description": "썸네일 이미지 URL (파일 업로드로 자동 처리되므로 일반적으로 사용하지 않음)",
            "example": "https://res.cloudinary.com/demo/image/upload/v1123456789/sample.jpg"
          }
        }
      },
...
      "CommentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "댓글 ID (자동생성)",
            "example": 1
          },
          "content": {
            "type": "string",
            "description": "댓글 내용",
            "example": "좋은 게시글이네요!"
          },
          "userId": {
            "type": "number",
            "description": "작성자 ID",
            "example": 1
          },
          "postId": {
            "type": "number",
            "description": "게시글 ID",
            "example": 1
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "댓글 생성 시간",
            "example": "2025-06-09T13:30:51+09:00"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "댓글 수정 시간",
            "example": "2025-06-09T13:30:51+09:00"
          },
          "user": {
            "type": "object",
            "description": "작성자 정보",
            "example": {
              "id": 1,
              "username": "사용자"
            }
          }
        },
        "required": [
          "id",
          "content",
          "userId",
          "postId",
          "createdAt",
          "updatedAt",
          "user"
        ]
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "댓글 내용",
            "example": "좋은 글이네요!",
            "maxLength": 1000
          }
        },
        "required": [
          "content"
        ]
      },
...
    }
  }
}